Updated to co-exist with sl_bt_event
