+++ src/platform/silabs/efr32/BLEChannel.h
@@ -23,30 +23,67 @@ struct BLEConState
     uint8_t bondingHandle;
 };
 
+struct AdvConfigStruct
+{
+    ByteSpan advData;
+    ByteSpan responseData;
+    uint32_t intervalMin;
+    uint32_t intervalMax;
+    uint16_t duration;
+    uint8_t maxEvents;
+};
+
 class BLEChannel
 {
 public:
-    BLEChannel();
+    BLEChannel()
+    {
+        // Initialize the connection state
+        memset(&mConnectionState, 0, sizeof(mConnectionState));
+        mFlags.ClearAll();
+        mAdvIntervalMin = 0;
+        mAdvIntervalMax = 0;
+        mAdvDuration    = 0;
+        mAdvMaxEvents   = 0;
+    }
+
     virtual ~BLEChannel() = default;
 
-    CHIP_ERROR Init(void);
+    virtual CHIP_ERROR Init(void) = 0;
 
     // TODO: Revisit this method and see about splitting timing vs advertising data. It is currently implemented to do everything
-    // the Matter BLEManager requires but it should be modified to use existing APIs for the advertising data and timing
-    CHIP_ERROR ConfigureAdvertising(ByteSpan advData, ByteSpan responseData, uint32_t intervalMin, uint32_t intervalMax,
-                                    uint16_t duration, uint8_t maxEvents);
+    /** @brief ConfigureAdvertising
+     *   Configure the advertising data and parameters for the BLE channel. This function needs to be called before
+     *  starting the advertising process.
+     *
+     *  @param AdvConfigStruct Structure holding The advertising data to be set, the scan response data to be set,
+     *  the minimum and maximum advertising intervals, the duration of the advertising, and the maximum number of events.
+     */
+    virtual CHIP_ERROR ConfigureAdvertising(const AdvConfigStruct & config) = 0;
 
     /** @brief StartAdvertising
      *  Start the advertising process for the BLE channel using configured parameters. ConfigureAdvertising must be called before
      * this function.
      */
-    CHIP_ERROR StartAdvertising(void);
-    CHIP_ERROR StopAdvertising(void);
+    virtual CHIP_ERROR StartAdvertising(void) = 0;
+    virtual CHIP_ERROR StopAdvertising(void)  = 0;
 
-    virtual void AddConnection(uint8_t connectionHandle, uint8_t bondingHandle);
-    virtual bool RemoveConnection(uint8_t connectionHandle);
-    virtual void HandleReadRequest(volatile sl_bt_msg_t * evt, ByteSpan data);
-    virtual void HandleWriteRequest(volatile sl_bt_msg_t * evt, MutableByteSpan data);
+    virtual void AddConnection(uint8_t connectionHandle, uint8_t bondingHandle)       = 0;
+    virtual bool RemoveConnection(uint8_t connectionHandle)                           = 0;
+    virtual void HandleReadRequest(volatile sl_bt_msg_t * evt, ByteSpan data)         = 0;
+    virtual void HandleWriteRequest(volatile sl_bt_msg_t * evt, MutableByteSpan data) = 0;
+
+    /** @brief CanHandleEvent
+     *  Determine if the side channel can handle a BLE event. This only gets called if the
+     *  targetted event is not supported by MATTERoBLE as the side channel supports every
+     *  event that MATTERoBLE does.
+     *
+     * @param event uint32_t Header of sl_bt_msg_t event type
+     * @return true if the event can be handled, false otherwise.
+     */
+    virtual bool CanHandleEvent(uint32_t event) { return false; }
+
+    virtual void ParseEvent(volatile sl_bt_msg_t * evt) {}
 
     /** @brief CCCD Write Handler
      *
@@ -61,16 +98,16 @@ public:
      * @return CHIP_ERROR_INCORRECT_STATE if a request is received when the connection is not allocated or a request is received
      *        for a different connection handle. CHIP_NO_ERROR if the request is handled successfully.
      */
-    virtual CHIP_ERROR HandleCCCDWriteRequest(volatile sl_bt_msg_t * evt, bool & isNewSubscription);
+    virtual CHIP_ERROR HandleCCCDWriteRequest(volatile sl_bt_msg_t * evt, bool & isNewSubscription) = 0;
 
-    virtual void UpdateMtu(volatile sl_bt_msg_t * evt);
+    virtual void UpdateMtu(volatile sl_bt_msg_t * evt) = 0;
 
     // CLI methods BEGIN
     // GAP
-    virtual CHIP_ERROR GeneratAdvertisingData(uint8_t discoverMove, uint8_t connectMode, const Optional<uint16_t> & maxEvents);
+    virtual CHIP_ERROR GeneratAdvertisingData(uint8_t discoverMove, uint8_t connectMode, const Optional<uint16_t> & maxEvents) = 0;
     virtual CHIP_ERROR SetAdvertisingParams(uint32_t intervalMin, uint32_t intervalMax, uint16_t duration,
                                             const Optional<uint16_t> & maxEvents, const Optional<uint8_t> & channelMap);
-    virtual CHIP_ERROR OpenConnection(bd_addr address, uint8_t addrType);
+    virtual CHIP_ERROR OpenConnection(bd_addr address, uint8_t addrType) = 0;
 
     /** @brief SetConnectionParams
      *
@@ -87,8 +124,8 @@ public:
      * otherwise.
      */
     virtual CHIP_ERROR SetConnectionParams(const Optional<uint8_t> & connectionHandle, uint32_t intervalMin, uint32_t intervalMax,
-                                           uint16_t latency, uint16_t timeout);
-    virtual CHIP_ERROR CloseConnection(void);
+                                           uint16_t latency, uint16_t timeout) = 0;
+    virtual CHIP_ERROR CloseConnection(void)                                   = 0;
 
     /** @brief Set Advertising Handle
      *
@@ -104,13 +141,13 @@ public:
      *  @param handle The advertising handle to set as the channel's advertising handle.
      *  @return CHIP_NO_ERROR if the handle is set successfully, otherwise an error code.
      */
-    virtual CHIP_ERROR SetAdvHandle(uint8_t handle);
+    virtual CHIP_ERROR SetAdvHandle(uint8_t handle) = 0;
 
     // GATT (All these methods need some event handling to be done in sl_bt_on_event)
-    virtual CHIP_ERROR DiscoverServices();
-    virtual CHIP_ERROR DiscoverCharacteristics(uint32_t serviceHandle);
-    virtual CHIP_ERROR SetCharacteristicNotification(uint8_t characteristicHandle, uint8_t flags);
-    virtual CHIP_ERROR SetCharacteristicValue(uint8_t characteristicHandle, const ByteSpan & value);
+    virtual CHIP_ERROR DiscoverServices()                                                           = 0;
+    virtual CHIP_ERROR DiscoverCharacteristics(uint32_t serviceHandle)                              = 0;
+    virtual CHIP_ERROR SetCharacteristicNotification(uint8_t characteristicHandle, uint8_t flags)   = 0;
+    virtual CHIP_ERROR SetCharacteristicValue(uint8_t characteristicHandle, const ByteSpan & value) = 0;
     // CLI methods END
 
     // Getters
@@ -120,7 +157,7 @@ public:
     bd_addr GetRandomizedAddr(void) const { return mRandomizedAddr; }
     BLEConState GetConnectionState() const { return mConnectionState; }
 
-private:
+protected:
     enum class Flags : uint16_t
     {
         kAdvertising = 0x0001, // Todo : See about flags for connection, subscription, etc.
